oh boy, that's a lot of moving parts
