+++ act/nari_act/__init__.py
@@ -1,0 +1,1 @@
+from nari_act.actlog import ActLogReader
