+++ act/nari_act/actlog.py
@@ -2,11 +2,11 @@
 from typing import Optional
 
 from nari.io.reader import Reader
-from nari.io.reader.actlogutils import ID_MAPPINGS, ActEventType, ActLogChecksumType, date_from_act_timestamp, validate_checksum
 from nari.types.event import Event
 from nari.types.event.version import SemanticVersion, Version
 from nari.util.exceptions import EventNotFound
-from nari.io.reader.actlogutils.exceptions import InvalidActChecksum
+from nari_act.actlogutils.exceptions import InvalidActChecksum
+from nari_act.actlogutils import ID_MAPPINGS, ActEventType, ActLogChecksumType, date_from_act_timestamp, validate_checksum
 
 # The last version of ACT to use MD5 checksums
 # We default to SHA256, but if the version is <= 2.2.1.6 we switch to MD5
