+++ act/nari_act/actlogutils/effectresult.py
